1601 CustomerFactory split
